feat: Rename sharePage component
